comment-rate-service/requirements.txt
django>=4.2,<5.3
djangorestframework>=3.14,<3.17
psycopg2-binary>=2.9
requests>=2.31
dj-database-url
django-filter
